+++ src/constants/helpers.tsx
@@ -55,7 +55,8 @@ export const termsAndConditionDocumentUploadMaxSize = {
 };
 
 export const repairJobProcessSteps = ['Serialized Assets', 'Repair Process'];
-export const salesOrderProcessSteps = ['Add Products', 'Add Services', 'Serialized Asset', 'Loading Ticket', 'Ready To Invoice'];
+//export const salesOrderProcessSteps = ['Add Products', 'Add Services', 'Serialized Asset', 'Loading Ticket', 'Ready To Invoice'];
+export const salesOrderProcessSteps = ['Add Products', 'Add Services', 'Ready To Invoice'];
 export const purchaseOrderSteps = ['Add Product', 'Services and Consumables', 'Issue', 'Receiving Asset'];
 export const rentalManagementSteps = [
   'Add Products',
@@ -389,8 +390,7 @@ export const repairJob = {
 };
 
 export const salesOrder = {
-  salesOrderResource: 'salesOrder',
-  salesOrderApi: '/sales-order',
+  api: '/sales-order',
   resource: 'sales-order'
 };